requirements.txt
discord.py
pandas
python-dotenv
seaborn
japanize_matplotlib
setuptools